EH-1732: if initial-state is based on opiskeluoikeus, handle later again
This is the change agreed on in the team about handling palautteet based
on handling date (käsittelypäivä) information.  Because opiskeluoikeus
data is external and can change without notification to eHOKS, we put
palautteet in :odottaa-kasittelya state when the verdict is based on
opiskeluoikeus fields.  This causes them to be handled on their
heratepvm (or later), using the information in Koski on that date.  This
means that the opiskeluoikeus information can be corrected during this
time.

When heratepvm has gone, the only way to retrigger palaute handling is
to resave the HOKS.  This usually makes sense because the only way to
change the heratepvm is also to resave.