EH-1840: delete old palautteet
This is mainly because they have been constructed during the time of
palaute-backend development and some of them should not be there if they
can already be found in herätepalvelu.  But it is complicated to mass
check whether the palautteet are already in herätepalvelu, so instead,
we purge everything and add back those that are not.